BUG: Propagate volume selection when slice views are linked
In the case when a slice view uniquely display a volume (as a background layer for example), turn on the "linked views" option, the displayed volumes don't change as expected, however, select the same volume in the slice view that already display the volume (as a background layer), the user would expect to have all 3 slices (red, green and yellow) display that volume (as a background layer). This is only possible using the recently added signal qMRMLNodeComboBox::nodeActivated. Caution should be used when changing the behavior.

git-svn-id: http://svn.slicer.org/Slicer4/trunk@15890 3bd1e089-480b-0410-8dfb-8563597acbee