COMP: Fix SlicerFunctionCheckCompilerFlags so that successive invocation works
Internally, the macro CMAKE_CXX_ACCEPTS_FLAG calls TRY_COMPILE. To avoid
the cost of compiling the test each time the project is configured, the variable set by
the macro is added to the cache so that following invocation of the macro with
the same variable name skip the compilation step.

For that same reason, SlicerFunctionCheckCompilerFlags function appends a unique suffix to
the HAS_FLAG variable. This suffix is created using a 'clean version' of the flag to test

See also https://github.com/commontk/CTK/commit/95b5aac71de69145aedcf40925d9daf59355f700

git-svn-id: http://svn.slicer.org/Slicer4/trunk@15999 3bd1e089-480b-0410-8dfb-8563597acbee