ENH: Add a work-in-progress DICOM module
This is based on the CTK dicom work, with some extra integration to make it
more compatible with the needs to slicer.  The goal is to get the 
basic interface and workflow figured out and then see what changes
should be made to CTK itself.

There is also the option of creating a 'mini PACS' for testing that can be used to query/retrieve data into the local database.


git-svn-id: http://svn.slicer.org/Slicer4/trunk@16750 3bd1e089-480b-0410-8dfb-8563597acbee