BUG: Don't calculate auto values in volume display node of vtkMRMLSlicerLayerLogic
The duplicate display node of vtkMRMLSliceLayerLogic shouldn't have to compute the scalar window/level or threshold as it it supposed to copy only values of the displaynode.

This was producing errors in vtkImageResliceMask because the extent of the background image mask (output 1) was different than the image (output 0). This is because CalculateAutoLevels is only plugged to output 0, not output1.

git-svn-id: http://svn.slicer.org/Slicer4/trunk@17841 3bd1e089-480b-0410-8dfb-8563597acbee