BUG: No need for vtkDiffusionTensorGlyph::Dimensions
Input dimensions can be automatically computed in vtkDiffusionTensorGlyph::RequestData. Due to some late update, the correct dimensions of the input image were set after RequestData was called with an empty Dimensions. Having no "Dimensions" resulted in computing the glyphs on the entire image (instead of a subsample using DimensionResolution).
The subsampling is now automatic. The loading of the scenes is then faster and require less memory.

Closes #1396
Closes #1407

git-svn-id: http://svn.slicer.org/Slicer4/trunk@18029 3bd1e089-480b-0410-8dfb-8563597acbee