ENH: remove ThreeDViews and SlicesControllerToolbar
These two collapsable frames at the bottom of the module panel
were slated for removal as part of the cosmetic overhaul leading
up to the RSNA release.  J2 moved all of the slice-related functionality
into the slice controllers and 3d functionality into the 3d controller
so there was no longer a need for these.

Nicole is working on moving the SceneView buttons to the main 
toolbar.

Also removed here are the slice magnifier and the thumbnail
view of the 3D view.  We (Ron and Steve) determined that these were 
not serving their original intention and were taking up space
that would be better served by the DataProbe information.
In the future we envision the ability to create a slice magnifier
view using the layout manager.



git-svn-id: http://svn.slicer.org/Slicer4/trunk@18030 3bd1e089-480b-0410-8dfb-8563597acbee