ENH: Add dependencies between modules
In order for modules to initialize their logics that depend on other module logics, the modules must be initialized in a specific order controlled by the graph of module dependencies.
The logics shouldn't have to access data from the application, if they need a logic, it should be set externally by the logic instantiator (in the case of vtkSlicerCropVolumeLogic it is qSlicerCropVolumeModule).
The CLI module logic still needs to be split away from the module (in a comming commit).

git-svn-id: http://svn.slicer.org/Slicer4/trunk@18978 3bd1e089-480b-0410-8dfb-8563597acbee