ENH: added controls to edit chart title and axis labels.
In ChartViewController, QLineEdit widgets are used for the title
and axis labels.  But QLineEdit cannot grab the keyboard focus when
inside the type of widget that makes up the ControllerWidget popup.
This may be fixed by changing the type of widget used for ControllerWidgets.
In the interim, added buttons to bring up a dialog for editing the
chart titles and labels.

git-svn-id: http://svn.slicer.org/Slicer4/trunk@19292 3bd1e089-480b-0410-8dfb-8563597acbee