BUG: fix case where dicom dialog gets stuck behind main window
Since the DICOM window was a dialog of the mainWindow and progress
dialogs from ctk and slicer are/were parented to the dicomApp
or mainWindow they there could be a race condition when the 
progress dialog was dismissed.  Also, users (Andras) did not like
the modal behavior of the dialog.  To fix these the window
is not a toplevel and slicer progress dialogs are relative to 
this.  This also allows the window to be kept open during 
other operations.



git-svn-id: http://svn.slicer.org/Slicer4/trunk@19638 3bd1e089-480b-0410-8dfb-8563597acbee