BUG: Ensure "local file" serverUrl is properly set in qSlicerExtensionsManagerModelTest
* For each test, the Midas server url is set to be a local file, that
way when the model manager attempt to query the server it gets the content
of the file.

For example, if the server url is set to be "file:///foo" and if the query
is "file:///foo/api/json?method=midas.version", the content of the file
"/foo/api/json" will correspond to the reply.

On windows, an extra "/" has to be added when specifying a local file as Url.
Since we were manually pre-pending "file://", it was working for unix-like
system but not for windows. See [1] for more details.

This commit make use of "QUrl::fromLocalFile" to address that issue. Note
also that the unit test has been updated.

[1] http://en.wikipedia.org/wiki/File_URI_scheme#Windows_2

git-svn-id: http://svn.slicer.org/Slicer4/trunk@19797 3bd1e089-480b-0410-8dfb-8563597acbee